requirements.txt
streamlit
gspread
streamlit-authenticator==0.2.2
matplotlib
plotly
dotenv
google.generativeai
langchain_community
langchain_google_genai
tabulate